# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CodeVault API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{id}/files:
    get:
      operationId: listProjectFiles
      tags: [repository]
      summary: List files in project repository
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RepoFile"
  /projects/{id}/commits:
    get:
      operationId: listCommits
      tags: [repository]
      summary: List commit history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Commits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Commit"
  /projects/{id}/branches:
    get:
      operationId: listBranches
      tags: [repository]
      summary: List branches
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Branches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Branch"
  /projects/{id}/issues:
    get:
      operationId: listIssues
      tags: [repository]
      summary: List issues
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Issues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Issue"
    post:
      operationId: createIssue
      tags: [repository]
      summary: Create an issue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueInput"
      responses:
        "201":
          description: Created issue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"

  # Deployments
...
  /users/{username}:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get user profile by username
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /trading/signals:
    get:
      operationId: listTradingSignals
      tags: [trading]
      summary: List trading signals
      responses:
        "200":
          description: Signals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TradingSignal"
    post:
      operationId: createTradingSignal
      tags: [trading]
      summary: Create a trading signal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradingSignalInput"
      responses:
        "201":
          description: Created signal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradingSignal"
...
components:
  schemas:
...
    RepoFile:
      type: object
      required: [name, path, type, size]
      properties:
        name:
          type: string
        path:
          type: string
        type:
          type: string
        size:
          type: integer
        lastCommit:
          type: ["string", "null"]
        lastCommitDate:
          type: ["string", "null"]
    Commit:
      type: object
      required: [id, sha, message, author, date]
      properties:
        id:
          type: integer
        sha:
          type: string
        message:
          type: string
        author:
          type: string
        date:
          type: string
        additions:
          type: integer
        deletions:
          type: integer
    Branch:
      type: object
      required: [name, isDefault, lastCommit, updatedAt]
      properties:
        name:
          type: string
        isDefault:
          type: boolean
        lastCommit:
          type: string
        updatedAt:
          type: string
        aheadBy:
          type: integer
        behindBy:
          type: integer
    Issue:
      type: object
      required: [id, projectId, title, status, priority, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        priority:
          type: string
        labels:
          type: array
          items:
            type: string
        createdAt:
          type: string
    IssueInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
        labels:
          type: array
          items:
            type: string
...
    UserProfile:
      type: object
      required: [id, username, displayName, bio, totalProjects, totalStars, followers, following, joinedAt]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        totalProjects:
          type: integer
        totalStars:
          type: integer
        followers:
          type: integer
        following:
          type: integer
        achievements:
          type: array
          items:
            type: string
        joinedAt:
          type: string
...
    TradingSignal:
      type: object
      required: [id, pair, direction, entryPrice, targetPrice, stopLoss, status, confidence, createdAt]
      properties:
        id:
          type: integer
        pair:
          type: string
        direction:
          type: string
        entryPrice:
          type: number
        targetPrice:
          type: number
        stopLoss:
          type: number
        status:
          type: string
        confidence:
          type: number
        pnl:
          type: ["number", "null"]
        createdAt:
          type: string
    TradingSignalInput:
      type: object
      required: [pair, direction, entryPrice, targetPrice, stopLoss, confidence]
      properties:
        pair:
          type: string
        direction:
          type: string
        entryPrice:
          type: number
        targetPrice:
          type: number
        stopLoss:
          type: number
        confidence:
          type: number
...