# lib/api-zod/src/generated/types/branch.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeVault API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Branch {
  name: string;
  isDefault: boolean;
  lastCommit: string;
  updatedAt: string;
  aheadBy?: number;
  behindBy?: number;
}