# lib/api-zod/src/generated/types/commit.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeVault API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Commit {
  id: number;
  sha: string;
  message: string;
  author: string;
  date: string;
  additions?: number;
  deletions?: number;
}