# lib/api-zod/src/generated/types/issue.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeVault API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Issue {
  id: number;
  projectId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: string;
  priority: string;
  labels?: string[];
  createdAt: string;
}