# lib/api-zod/src/generated/types/repoFile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeVault API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RepoFile {
  name: string;
  path: string;
  type: string;
  size: number;
  /** @nullable */
  lastCommit?: string | null;
  /** @nullable */
  lastCommitDate?: string | null;
}