# lib/api-zod/src/generated/types/tradingSignal.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeVault API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TradingSignal {
  id: number;
  pair: string;
  direction: string;
  entryPrice: number;
  targetPrice: number;
  stopLoss: number;
  status: string;
  confidence: number;
  /** @nullable */
  pnl?: number | null;
  createdAt: string;
}